DRY teacher code using parent class Teacher
