+++ apprentice_teacher.rb
@@ -1,46 +1,18 @@
 require_relative 'dbc_member'
+require_relative 'teacher'
 
-class ApprenticeTeacher < DBCMember
+class ApprenticeTeacher < Teacher
   attr_reader :salary, :target_raise
+
+  TEACH_STUFF_LINE_1 = "Listen, class, this is how everything works. "
+  TEACH_STUFF_LINE_2 = "*drops crazy knowledge bomb* "
+  TEACH_STUFF_LINE_3 = "... You're welcome."
+
+  RATING_THRESHOLD = 80
 
   def initialize(options={})
     super
     @target_raise = 800
-    @phase = 3
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
-
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works. "
-    response += "*drops crazy knowledge bomb* "
-    response += "... You're welcome."
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 80
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
   end
 
   def attend_training_session
