+++ senior_teacher.rb
@@ -1,46 +1,18 @@
 require_relative 'dbc_member'
+require_relative 'teacher'
 
-class SeniorTeacher < DBCMember
+class SeniorTeacher < Teacher
   attr_reader :salary, :performance_rating, :target_raise
+
+  TEACH_STUFF_LINE_1 = "Listen, class, this is how everything works, fo SHO! "
+  TEACH_STUFF_LINE_2 = "*drops flat-out insane knowledge bomb* "
+  TEACH_STUFF_LINE_3 = "... You're welcome. *saunters away*"
+
+  RATING_THRESHOLD = 90
 
   def initialize(options={})
     super
-    @phase = 3
     @target_raise = 1000
-  end
-
-  def set_phase(num)
-    @phase = num
-    "Cool, I've always wanted to teach phase #{num}!"
-  end
-
-  def teach_stuff
-    response = ""
-    response += "Listen, class, this is how everything works, fo SHO! "
-    response += "*drops flat-out insane knowledge bomb* "
-    response += "... You're welcome. *saunters away*"
-    response
-  end
-
-  def salary=(new_salary)
-    puts "This better be good!"
-    @salary = new_salary
-  end
-
-  def receive_raise(raise)
-    @salary += raise
-  end
-
-  def set_performance_rating(rating)
-    response = ""
-    if rating > 90
-      response = "Yay, I'm a great employee!"
-      receive_raise(@target_raise)
-    else
-      response += "Oh, well -- thanks to this actionable, specific, and kind "
-      response += "feedback, I'll do better next time."
-    end
-    response
   end
 
   def lead_training_session
