+++ teacher.rb
@@ -1,0 +1,44 @@
+require_relative 'dbc_member'
+
+class Teacher < DBCMember
+
+  def initialize(options = {})
+    super
+    @phase = 3
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+  def set_phase(num)
+    @phase = num
+    "Cool, I've always wanted to teach phase #{num}!"
+  end
+
+  def teach_stuff
+    response = ""
+    response += self.class::TEACH_STUFF_LINE_1
+    response += self.class::TEACH_STUFF_LINE_2
+    response += self.class::TEACH_STUFF_LINE_3
+    response
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if rating > self.class::RATING_THRESHOLD
+      response = "Yay, I'm a great employee!"
+      receive_raise(@target_raise)
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += "feedback, I'll do better next time."
+    end
+    response
+  end
+
+end
